fix issue where touch was interpreted as move
In practice, I could only expand the channels section when using
the emulator and being careful not to use my mouse. On a real
device, I could not expand the channels section at all. The
solution was to put in a small acceptable error sensitivity that
caused the view to not interpret all touches as moves.